+++ lib/f_service.rb
@@ -6,4 +6,12 @@ require_relative 'f_service/base'
 #
 # @api public
 module FService
+  # Marks a method as deprecated
+  #
+  # @api private
+  def self.deprecate!(name:, alternative:)
+    warn "[DEPRECATED] #{name} is deprecated; " \
+         "use #{alternative} instead. " \
+         'It will be removed on the next release.'
+  end
 end